requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
fastapi-utils==0.2.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
